+++ spec/telegram/bot/client_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Telegram::Bot::Client do
   let(:botan_token) { double(:botan_token) }
 
   include_examples 'initializers'
-  include_examples 'async', request_args: -> { [double(:action), {body: :content}] }
+  it_behaves_like 'async', request_args: -> { [double(:action), {body: :content}] }
 
   describe '.prepare_body' do
     subject { described_class.prepare_body(input) }
@@ -46,6 +46,51 @@ RSpec.describe Telegram::Bot::Client do
       its(:token) { should eq args[0][:token] }
       its(:username) { should eq args[0][:username] }
       its(:base_uri) { should include args[0][:token] }
+    end
+  end
+
+  describe '#request' do
+    subject { -> { instance.request(action, request_body) } }
+    let(:action) { :some_action }
+    let(:url) { "#{format described_class::URL_TEMPLATE, token}#{action}" }
+    let(:request_body) { double(:body) }
+    let(:prepared_body) { double(:prepared_body) }
+    let(:response) { HTTP::Message.new_response(body).tap { |x| x.status = status } }
+
+    let(:status) { 200 }
+    let(:body) { body_json.to_json }
+    let(:body_json) { {'param' => 'val', 'description' => 'some description'} }
+    before do
+      expect(described_class).to receive(:prepare_body).with(request_body) { prepared_body }
+      expect(instance).to receive(:http_request).with(url, prepared_body) { response }
+    end
+    around { |ex| Telegram::Bot::ClientStub.stub_all!(false) { ex.run } }
+
+    shared_examples 'invalid body' do |error = Telegram::Bot::Error|
+      context 'when body is not json' do
+        let(:body) { '{' }
+        it { should raise_error error }
+      end
+    end
+
+    its(:call) { should eq body_json }
+    include_examples 'invalid body', JSON::ParserError
+
+    context 'when status is 403' do
+      let(:status) { 403 }
+      it { should raise_error Telegram::Bot::Forbidden, body_json['description'] }
+      include_examples 'invalid body'
+    end
+
+    context 'when status is 404' do
+      let(:status) { 404 }
+      it { should raise_error Telegram::Bot::NotFound, body_json['description'] }
+      include_examples 'invalid body'
+    end
+
+    context 'when status is other' do
+      let(:status) { 500 }
+      it { should raise_error Telegram::Bot::Error, /#{body_json['description']}/ }
     end
   end
 end